requirements.txt
setuptools~=60.2.0
nox~=2022.1.7
py-multihash~=2.0.1
